Remove sharding for repackage step
Sharding for the repackage step is broken at least in the following two
ways:
- The repackage target expects all packages for one run to be available,
  sharding breaks this.
- Packages meant to be used in a single bundle could land on separate
  shards.

On top of that, we usually only crawl a couple of new packages at a
time and even when adding support for a new distro, we should only get
up to a couple hundred of packages, which translates to a couple GB of
disk space needed, so GH runners should have enough disk space even in
this condition.

Cache invalidation of the entire set of bundles and packages is also not
a thing we'll ever do, since the time taken to reprocess the entire
thing would require an amount of resources that no CI system will
provide for us. If we need to do something like this, we'll be better
off removing everything and slowly readding distros to kernel-crawler.py